x86-64 inline c syscall: no 64 bit cast

+++ x86-64/c/linux/hello.c
@@ -6,8 +6,8 @@ ssize_t my_write(int fd, const void *buf, size_t size) {
     ssize_t ret;
     __asm__ __volatile__ (
         "syscall"
-        : "=a" ((int64_t)ret)
-        : "0"(1), "D"(fd), "S"((int64_t)buf), "d"((uint64_t)size)
+        : "=a" (ret)
+        : "0" (1), "D" (fd), "S" (buf), "d" (size)
         : "cc", "rcx", "r11", "memory"
     );
     return ret;
@@ -17,8 +17,8 @@ void my_exit(int exit_status) {
     ssize_t ret;
     __asm__ __volatile__ (
         "syscall"
-        : "=a" ((int64_t)ret)
-        : "0" ((int64_t)60), "D" ((int64_t)exit_status)
+        : "=a" (ret)
+        : "0" (60), "D" (exit_status)
         : "cc", "rcx", "r11", "memory"
     );
 }
